New save and sample
This currently is tested to work for both unconditional and upsampler. However, the coloring the original images for the upsampler look off. I'm not sure what the issue is, so please have a look.

Note that this moves the generator kwargs arguments into a separate definition in the GigaGan class so it can be called in the save and sample, too.